requirements.txt
Django==4.1.2
django-ordered-model==3.6
djangorestframework==3.14.0
